+++ src/_utils/story.tsx
@@ -4,7 +4,7 @@ import { withKnobs, text } from '@storybook/addon-knobs'
 
 import { color, space, font } from '_utils/branding'
 
-const stories = storiesOf('Brand', module)
+const stories = storiesOf('Widgets|Brand', module)
 stories.addDecorator(withKnobs)
 
 const styles: { [name: string]: React.CSSProperties } = {