+++ src/theVoice/story.tsx
@@ -5,7 +5,7 @@ import { withKnobs } from '@storybook/addon-knobs'
 import TheVoice from '.'
 import readme from 'theVoice/specifications/theVoice.md'
 
-const stories = storiesOf('TheVoice', module)
+const stories = storiesOf('Widgets|TheVoice', module)
 stories.addDecorator(withKnobs)
 
 stories.add('Specifications', () => null, {